tests/wpt/web-platform-tests/trusted-types/location-href.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>
<body>
<script>
  promise_test(t => {
    var url = TrustedURL.create(URLS.safe);
    return createFrameAndHref(url);
  }, "Basic processing: safe URL, safe construction.");

  promise_test(t => {
    var url = TrustedURL.unsafelyCreate(URLS.safe);
    return createFrameAndHref(url);
  }, "Basic processing: safe URL, unsafe construction.");

  promise_test(t => {
    var url = TrustedURL.create(URLS.javascript);
    return createFrameAndHref(url).then(i => {
      assert_equals(i.contentDocument.body.innerText, "");
    });
  }, "Basic processing: javascript URL, safe construction.");

  promise_test(t => {
    var url = TrustedURL.unsafelyCreate(URLS.javascript);
    return createFrameAndHref(url).then(i => {
      assert_equals(i.contentDocument.body.innerText, "scripted");
    });
  }, "Basic processing: javascript URL, unsafe construction.");
</script>
